Add `mapPalette` URL parameter
The mapping between the palette and the temperature anomalies can either
be computed on a per-region basis (`per-region`, default) or for a
single region (index) that is then applied to the whole dataset. Using a
single dataset as a reference makes it easier to compare temperature
anomalies of different regions by color. Otherwise, the same temperature
anomaly might be displayed in different colors for different regions.